feat(layout) App clean up and Layout creation
