+++ src/app/app-routing.module.ts
@@ -1,18 +1,29 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from './guards/auth.guard';
+import { LayoutComponent } from './layout/layout.component';
 
 const routes: Routes = [
   {
-    path: 'auth',
-    loadChildren: () => import('@app-auth/auth.module').then(m => m.AuthModule)
+    path: '',
+    redirectTo: '/auth/welcome',
+    pathMatch: 'full'
   },
   {
-    path: 'home',
-    loadChildren: () => import('./home/home.module').then(m => m.HomeModule),
-    canLoad: [AuthGuard]
+    path: '',
+    component: LayoutComponent,
+    children: [
+      {
+        path: 'auth',
+        loadChildren: () => import('@app-auth/auth.module').then(m => m.AuthModule)
+      },
+      {
+        path: 'home',
+        loadChildren: () => import('./home/home.module').then(m => m.HomeModule),
+        canLoad: [AuthGuard]
+      },
+    ]
   },
-  { path: '', redirectTo: 'auth', pathMatch: 'full' },
   { path: '**', redirectTo: 'auth' },
 ];
