+++ src/app/auth/login/login.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnDestroy } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Subject } from 'rxjs';
@@ -10,7 +10,7 @@ import { AuthService } from '@app-services/auth.service';
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss'],
 })
-export class LoginComponent implements OnDestroy {
+export class LoginComponent implements OnDestroy, OnInit {
   public loginForm: FormGroup;
   private destroy$: Subject<boolean> = new Subject<boolean>();
 
@@ -61,7 +61,7 @@ export class LoginComponent implements OnDestroy {
 
   constructor(
     private router: Router,
+    private formBuilder: FormBuilder,
     private authService: AuthService,
-    private formBuilder: FormBuilder
-  ) {}
+  ) { }
 }
